Try to fix MappingParameter.MakeSane for `--polymerase`
`--polymerase` must work together with `--noSplitSubreads`,
and when both `--polymerase` and `--noSpltiSubreads` are set,
variable `polymerase` is True while variable `mapReadsSeparately`
is False.

Also USEPBBAM is used at compile time so that compiler knows if
pbbam library will be used to build blasr (as oppose to an
alternative way of compiling blasr without using pbbam library).